requirements.txt
acme==1.1.0
appdirs==1.4.4
attrs==19.3.0
automat==0.8.0
blinker==1.4
certbot==0.40.0
certifi==2019.11.28
chardet==3.0.4
click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configargparse==0.13.0
configobj==5.0.6
constantly==15.1.0
ctop==1.0.0
dbus-python==1.2.16
distlib==0.3.1
distro==1.4.0
distro-info==0.23ubuntu1
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-jenkins==0.110.0
domain-connect-dyndns==0.0.8
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jinja2==2.10.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
language-selector==0.1
launchpadlib==1.10.13
markupsafe==1.1.0
mock==3.0.5
more-itertools==4.2.0
netifaces==0.10.4
parsedatetime==2.4
pbr==5.4.5
pep8==1.7.1
pillow==7.2.0
pip-chill==1.0.0
psycopg2==2.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyflakes==2.2.0
pygobject==3.36.0
pyhamcrest==1.9.0
pyicu==2.4.2
pymacaroons==0.13.0
pynacl==1.3.0
pyopenssl==19.0.0
pyrfc3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-debian==0.1.36ubuntu1
pyyaml==5.3.1
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
social-auth-app-django==4.0.0
ssh-import-id==5.10
systemd-python==234
twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
